+++ src/main/java/us/pixelmemory/kevin/sdr/resamplers/AccumulatorBase.java
@@ -5,9 +5,9 @@ import us.pixelmemory.kevin.sdr.firfilters.LanczosTable;
 abstract class AccumulatorBase {
 	protected final LanczosTable lanczos;
 	protected float inPosition;
-	protected final double step;
+	protected final float step;
 
-	public AccumulatorBase(final LanczosTable lanczos, final float inPosition, final double inToOut) {
+	public AccumulatorBase(final LanczosTable lanczos, final float inPosition, final float inToOut) {
 		this.lanczos = lanczos;
 		this.inPosition = inPosition;
 		this.step = 1 / inToOut;
